# Reduce cyclocomp of `lint()` to <= 15