--- a/.lintr
+++ b/.lintr
@@ -1,7 +1,6 @@
-linters: linters_with_defaults( # The following TODOs are part of an effort to have {lintr} lint-free (#584)
-   line_length_linter(120),
+linters: linters_with_defaults(
+   line_length_linter(120)
    # TODO add implicit_integer_linter()
-   cyclocomp_linter(29) # TODO reduce to 15
  )
 exclusions: list(
   "inst/doc/creating_linters.R" = 1,