--- a/R/lint.R
+++ b/R/lint.R
@@ -629,7 +629,7 @@ maybe_report_progress <- function(done = FALSE) {
 
 maybe_append_error_lint <- function(lints, error, lint_cache, filename) {
   if (inherits(error, "lint")) {
-    lints <- append(lints, error)
+    lints[[length(lints) + 1L]] <- error
 
     if (!is.null(lint_cache)) {
       cache_lint(lint_cache, list(filename = filename, content = ""), "error", error)