Closes #1186