@AshesITR 2022-05-25
Okay, we're there. PTAL.
If it looks good, I'll do a quick performance comparison - the test suite runtime hasn't changed much, but I'll check on `tools/QC.R` and `default_linter_testcode.R` tomorrow if you think the code looks good.