@MichaelChirico 2022-05-25
Reading more carefully, I think it's fine. The copy is only needed in case of an error; I'm OK with prioritizing readability over pure efficiency in error regions.