@AshesITR 2022-05-26
Performance is unaffected for `default_linter_testcode.R`:
```
> bench::mark(lintr::lint("tests/testthat/default_linter_testcode.R", linters = lintr::linters_with_tags(NULL), parse_settings = FALSE), min_iterations = 5)

# main
  min   median `itr/sec` mem_alloc `gc/sec` n_itr  n_gc total_time result  memory time           gc      
  <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl> <int> <dbl>   <bch:tm> <list>  <list> <list>         <list>  
1 730ms    758ms      1.21        NA     13.0     5    54      4.15s <lints> <NULL> <bench_tm [5]> <tibble>

# this PR
  min   median `itr/sec` mem_alloc `gc/sec` n_itr  n_gc total_time result  memory time           gc      
  <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl> <int> <dbl>   <bch:tm> <list>  <list> <list>         <list>  
1 726ms    749ms      1.23        NA     14.5     5    59      4.08s <lints> <NULL> <bench_tm [5]> **<tibble**
```

In `tools/QC.R` I found a bug in `sprintf_linter()` for multi-line sprintf() calls with comments. #1249 
```
> bench::mark(
+   lintr::lint(test_file, linters = lintr::linters_with_tags(NULL, sprintf_linter = NULL), parse_settings = FALSE),
+   min_iterations = 5
+ )
# main
  min   median `itr/sec` mem_alloc `gc/sec` n_itr  n_gc total_time result  memory time           gc              
  <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl> <int> <dbl>   <bch:tm> <list>  <list> <list>         <list>          
1 36.2s    36.8s    0.0271        NA     6.83     5  1259      3.07m <lints> <NULL> <bench_tm [5]> <tibble [5 × 3]>

# this PR
  min   median `itr/sec` mem_alloc `gc/sec` n_itr  n_gc total_time result  memory time           gc              
  <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl> <int> <dbl>   <bch:tm> <list>  <list> <list>         <list>          
1 37.4s    37.6s    0.0264        NA     7.23     5  1367      3.15m <lints> <NULL> <bench_tm [5]> <tibble [5 × 3]>
```

LGTM, so merging.